# filmApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Film API",
        "description": "API de gestion des films, acteurs, critiques et utilisateurs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/films/{id}/actors": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Retourne les acteurs d'un film",
                "operationId": "6585262d4bb8eabaa2c2329ca7a21217",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des acteurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "PENELOPE"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "GUINESS"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film introuvable"
                    }
                }
            }
        },
        "/films/{id}/with-critics": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Retourne un film avec ses critiques",
                "operationId": "64df4d038d8d093fe933f03f5a0952c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film avec critiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "ACADEMY DINOSAUR"
                                        },
                                        "release_year": {
                                            "type": "integer",
                                            "example": 2006
                                        },
                                        "length": {
                                            "type": "integer",
                                            "example": 86
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "A Epic Drama of a Feminist..."
                                        },
                                        "rating": {
                                            "type": "string",
                                            "example": "PG"
                                        },
                                        "language_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "special_features": {
                                            "type": "string",
                                            "example": "Deleted Scenes,Behind the Scenes"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "",
                                            "nullable": true
                                        },
                                        "critics": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "film_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "score": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "example": "Très bon film"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film introuvable"
                    }
                }
            }
        },
        "/films/{film}/average-score": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Obtenir la moyenne des scores d'un film",
                "description": "Retourne la moyenne des scores pour un film particulier basé sur ses critiques.",
                "operationId": "852a708d585bf0301b7cdd4bdf2d7fae",
                "parameters": [
                    {
                        "name": "film",
                        "in": "path",
                        "description": "ID du film",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moyenne calculée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "film_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "average_score": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 7.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film introuvable"
                    }
                }
            }
        },
        "/films/search": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Recherche de films",
                "description": "Recherche des films par titre, description ou autres critères.",
                "operationId": "dab1b0610115e7958c15f5ce057bc479",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Terme de recherche",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "DINOSAUR"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de recherche paginés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "ACADEMY DINOSAUR"
                                                    },
                                                    "release_year": {
                                                        "type": "integer",
                                                        "example": 2006
                                                    },
                                                    "length": {
                                                        "type": "integer",
                                                        "example": 86
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "A Epic Drama of a Feminist..."
                                                    },
                                                    "rating": {
                                                        "type": "string",
                                                        "example": "PG"
                                                    },
                                                    "language_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "special_features": {
                                                        "type": "string",
                                                        "example": "Deleted Scenes,Behind the Scenes"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 72
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}